+++ src/components/formControlSection/__tests__/FormControlSection.Test.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
-import Input from 'src/components/input';
+import { Input } from 'src/components/input';
 
 import { render, screen } from 'jest-utils';
 
-import FormControlSection, { FormControlSectionProps } from '..';
+import { FormControlSection, FormControlSectionProps } from '..';
 
 const renderWrapper = ({
   id = 'test',
