+++ src/components/formControlSection/index.tsx
@@ -1,43 +1,37 @@
 import React, { FC, PropsWithChildren } from 'react';
-import {
-  Box,
-  FormControl as ChakraFormControl,
-  Flex,
-  FormErrorMessage,
-} from '@chakra-ui/react';
+import { Field as ChakraField } from '@chakra-ui/react';
 
-import Tooltip from '../tooltip';
-import Text from '../text';
-
+import { Tooltip } from '../tooltip';
+import { Text } from '../text';
 import { InformationIcon } from '../icons';
+import { Flex } from '../flex';
+import { Box } from '../box';
 
-export type Props = {
+export type BaseProps = {
   id: string;
   errorMessage?: string;
   label?: string;
   hint?: string;
   tooltip?: string;
-  recurring?: boolean;
 };
 
-const FormControlSection: FC<PropsWithChildren<Props>> = ({
+export const FormControlSection: FC<PropsWithChildren<BaseProps>> = ({
   children,
   errorMessage,
   id,
   label,
   hint,
   tooltip,
-  recurring,
 }) => {
   const isInvalid = !!errorMessage;
 
   return (
-    <ChakraFormControl id={id} isInvalid={isInvalid}>
+    <ChakraField.Root id={id} invalid={isInvalid}>
       <Box
-        border={recurring ? 'none' : '1px solid'}
+        border="1px"
         borderRadius="sm"
         borderColor={isInvalid ? 'red.500' : 'gray.400'}
-        padding={recurring ? '0' : '2xl'}
+        padding="2xl"
       >
         <Flex flexDirection="column">
           <Box mb="lg">
@@ -62,12 +56,9 @@ const FormControlSection: FC<PropsWithChildren<Props>> = ({
           {children}
         </Flex>
       </Box>
-      <FormErrorMessage>{errorMessage}</FormErrorMessage>
-    </ChakraFormControl>
+      <ChakraField.ErrorText>{errorMessage}</ChakraField.ErrorText>
+    </ChakraField.Root>
   );
 };
 
-FormControlSection.displayName = 'FormControlSection';
-
-export default FormControlSection;
-export { Props as FormControlSectionProps };
+export { BaseProps as FormControlSectionProps };
