+++ src/components/index.ts
@@ -46,3 +46,4 @@ export * from './themeProvider';
 export * from './vehicleReference';
 export * from './rangeFilterInput';
 export * from './topVehicleSharedBadge';
+export * from './formControlSection';
